'size_t' --> 'unsigned int'
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@35467 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 32b2f70a8a7a6ee1498a71f9600aedf79abe7058

+++ LHCb/Event/EventAssoc/dict/selection.h
@@ -1,6 +1,6 @@
-// $Id: selection.h,v 1.8 2006-11-25 19:01:15 ibelyaev Exp $
+// $Id: selection.h,v 1.9 2006-11-27 17:01:02 ibelyaev Exp $
 // ====================================================================
-// CVS tag $Name: not supported by cvs2svn $, version $Revision: 1.8 $ 
+// CVS tag $Name: not supported by cvs2svn $, version $Revision: 1.9 $ 
 // ====================================================================
 // Incldue files
 // ====================================================================
@@ -38,7 +38,7 @@
   *       
   * Helper file for building Reflex dictionaries for Relations 
    * @author ibelyaev 
-  * @date  2006-11-25 19:29:48.355258 
+  * @date  2006-11-27 17:50:34.699791 
   *       
   */      
 
@@ -59,13 +59,13 @@ GaudiDict::RelationWeighted1DDict<LHCb::CaloDigit,LHCb::MCParticle,float>	_10 ;
 GaudiDict::RelationWeighted1DDict<LHCb::ProtoParticle,LHCb::MCParticle,double>	_11 ;
 GaudiDict::RelationWeighted2DDict<LHCb::CaloCluster,LHCb::Track,float>	_12 ;
 GaudiDict::RelationWeighted2DDict<LHCb::CaloHypo,LHCb::Track,float>	_13 ;
-GaudiDict::RelationWeighted2DDict<LHCb::GenCollision,LHCb::RecVertex,std::pair<size_t,float> >	_14 ;
+GaudiDict::RelationWeighted2DDict<LHCb::GenCollision,LHCb::RecVertex,std::pair<unsigned int,float> >	_14 ;
 GaudiDict::RelationWeighted2DDict<LHCb::MCParticle,LHCb::Particle,double>	_15 ;
 GaudiDict::RelationWeighted2DDict<LHCb::MCParticle,LHCb::Track,double>	_16 ;
-GaudiDict::RelationWeighted2DDict<LHCb::MCVertex,LHCb::RecVertex,std::pair<size_t,float> >	_17 ;
+GaudiDict::RelationWeighted2DDict<LHCb::MCVertex,LHCb::RecVertex,std::pair<unsigned int,float> >	_17 ;
 GaudiDict::RelationWeighted2DDict<LHCb::Particle,LHCb::MCParticle,double>	_18 ;
-GaudiDict::RelationWeighted2DDict<LHCb::RecVertex,LHCb::GenCollision,std::pair<size_t,float> >	_19 ;
-GaudiDict::RelationWeighted2DDict<LHCb::RecVertex,LHCb::MCVertex,std::pair<size_t,float> >	_20 ;
+GaudiDict::RelationWeighted2DDict<LHCb::RecVertex,LHCb::GenCollision,std::pair<unsigned int,float> >	_19 ;
+GaudiDict::RelationWeighted2DDict<LHCb::RecVertex,LHCb::MCVertex,std::pair<unsigned int,float> >	_20 ;
 GaudiDict::RelationWeighted2DDict<LHCb::Track,LHCb::CaloCluster,float>	_21 ;
 GaudiDict::RelationWeighted2DDict<LHCb::Track,LHCb::CaloHypo,float>	_22 ;
 GaudiDict::RelationWeighted2DDict<LHCb::Track,LHCb::MCParticle,double>	_23 ;
